requirements.txt
flask
torch
torchvision
open-clip-torch
numpy
pandas
pillow
scikit-learn